lint and circular deps
