+++ src/types/event-types.ts
@@ -1,6 +1,5 @@
-import type { GenericMessageReply } from '../core/message-reply.js';
 import type { ProtocolsQueryFilter } from './protocols-types.js';
-import type { AuthorizationModel, GenericMessage } from './message-types.js';
+import type { AuthorizationModel, GenericMessage, GenericMessageReply } from './message-types.js';
 import type { DwnInterfaceName, DwnMethodName } from '../enums/dwn-interface-method.js';
 import type { RangeCriterion, RangeFilter } from './query-types.js';
