+++ src/types/messages-types.ts
@@ -1,5 +1,4 @@
-import type { GenericMessageReply } from '../core/message-reply.js';
-import type { AuthorizationModel, GenericMessage } from './message-types.js';
+import type { AuthorizationModel, GenericMessage, GenericMessageReply } from './message-types.js';
 import type { DwnInterfaceName, DwnMethodName } from '../enums/dwn-interface-method.js';
 
 export type MessagesGetDescriptor = {
