+++ src/types/method-handler.ts
@@ -1,6 +1,5 @@
-import type { GenericMessage } from './message-types.js';
-import type { GenericMessageReply } from '../core/message-reply.js';
 import type { Readable } from 'readable-stream';
+import type { GenericMessage, GenericMessageReply } from './message-types.js';
 
 /**
  * Interface that defines a message handler of a specific method.
