+++ src/types/protocols-types.ts
@@ -1,6 +1,5 @@
-import type { GenericMessageReply } from '../core/message-reply.js';
 import type { PublicJwk } from './jose-types.js';
-import type { AuthorizationModel, GenericMessage } from './message-types.js';
+import type { AuthorizationModel, GenericMessage, GenericMessageReply } from './message-types.js';
 import type { DwnInterfaceName, DwnMethodName } from '../enums/dwn-interface-method.js';
 
 export type ProtocolsConfigureDescriptor = {
